requirements.txt
python-telegram-bot==21.7
flask
google-generativeai
python-dotenv
gunicorn
